+++ data/trx/ship/games/tr3/scripts/house.lua
@@ -6,7 +6,7 @@ trx.events.before_item_setup(function(level)
   trx.items[59].properties.range = { x = 18, y = 2, z = 10 }
   trx.items[60].properties.range = { x = 26, y = 2, z = 10 }
   trx.items[78].properties.range = { x = 10, y = 2, z = 18 }
-  trx.items[57].properties.use_default_uv = false
-  trx.items[58].properties.use_default_uv = false
-  trx.items[60].properties.use_default_uv = false
+  trx.items[57].properties.sprite_offset = 1
+  trx.items[58].properties.sprite_offset = 1
+  trx.items[60].properties.sprite_offset = 1
 end)